refactor!: move legacy ui to /legacy
