+++ src/bedrock_server_manager/api/server_install_config.py
@@ -687,7 +687,7 @@ def install_new_server(
             "status": "success",
             "version": server.get_version(),
             "message": f"Server '{server_name}' installed successfully to version {server.get_version()}.",
-            "next_step_url": f"/server/{server_name}/configure_properties?new_install=true",
+            "next_step_url": f"/legacy/server/{server_name}/configure_properties?new_install=true",
         }
 
     except BSMError as e:
