+++ src/bedrock_server_manager/web/routers/setup.py
@@ -9,9 +9,8 @@ This module provides endpoints for:
 
 import logging
 
-from fastapi import APIRouter, Depends, HTTPException, Request, status
-from fastapi.responses import HTMLResponse, JSONResponse, RedirectResponse
-from fastapi.templating import Jinja2Templates
+from fastapi import APIRouter, Depends, HTTPException, status
+from fastapi.responses import JSONResponse
 from pydantic import BaseModel
 from sqlalchemy.exc import IntegrityError
 
@@ -19,11 +18,9 @@ from ...context import AppContext
 from ...db.models import User
 from ..auth_utils import (
     create_access_token,
-    get_current_user_optional,
     get_password_hash,
 )
-from ..dependencies import get_app_context, get_templates
-from ..schemas import User as UserSchema
+from ..dependencies import get_app_context
 
 logger = logging.getLogger(__name__)
 
@@ -43,25 +40,6 @@ async def get_setup_status(
     with app_context.db.session_manager() as db:  # type: ignore
         user_exists = db.query(User).first() is not None
         return {"needs_setup": not user_exists}
-
-
-@router.get("", response_class=HTMLResponse, include_in_schema=False)
-async def setup_page(
-    request: Request,
-    current_user: UserSchema = Depends(get_current_user_optional),
-    app_context: AppContext = Depends(get_app_context),
-    templates: Jinja2Templates = Depends(get_templates),
-):
-    """
-    Serves the setup page if no users exist in the database.
-    """
-    with app_context.db.session_manager() as db:  # type: ignore
-        if db.query(User).first():
-            # If a user already exists, redirect to home page, as setup is complete
-            return RedirectResponse(url="/", status_code=status.HTTP_302_FOUND)
-    return templates.TemplateResponse(
-        request, "setup.html", {"current_user": current_user}
-    )
 
 
 class CreateFirstUserRequest(BaseModel):
@@ -121,7 +99,7 @@ async def create_first_user(
                 content={
                     "status": "success",
                     "message": "Admin account created and logged in successfully.",
-                    "redirect_url": "/settings?in_setup=true",
+                    "redirect_url": "/legacy/settings?in_setup=true",
                 },
                 status_code=status.HTTP_200_OK,
             )
