+++ tests/web/routers/test_server_install_config_router.py
@@ -405,8 +405,9 @@ def test_install_server_api_route_invalid_name(
 
 def test_configure_properties_page(authenticated_client, real_bedrock_server):
     """Test the configure_properties_page route with a successful response."""
+    # Updated to point to legacy route
     response = authenticated_client.get(
-        f"/server/{real_bedrock_server.server_name}/configure_properties"
+        f"/legacy/server/{real_bedrock_server.server_name}/configure_properties"
     )
     assert response.status_code == 200
     assert "Server Properties" in response.text
@@ -414,8 +415,9 @@ def test_configure_properties_page(authenticated_client, real_bedrock_server):
 
 def test_configure_allowlist_page(authenticated_client, real_bedrock_server):
     """Test the configure_allowlist_page route with a successful response."""
+    # Updated to point to legacy route
     response = authenticated_client.get(
-        f"/server/{real_bedrock_server.server_name}/configure_allowlist"
+        f"/legacy/server/{real_bedrock_server.server_name}/configure_allowlist"
     )
     assert response.status_code == 200
     assert "Allowlist" in response.text
@@ -423,8 +425,9 @@ def test_configure_allowlist_page(authenticated_client, real_bedrock_server):
 
 def test_configure_permissions_page(authenticated_client, real_bedrock_server):
     """Test the configure_permissions_page route with a successful response."""
+    # Updated to point to legacy route
     response = authenticated_client.get(
-        f"/server/{real_bedrock_server.server_name}/configure_permissions"
+        f"/legacy/server/{real_bedrock_server.server_name}/configure_permissions"
     )
     assert response.status_code == 200
     assert "Permissions" in response.text
@@ -432,8 +435,9 @@ def test_configure_permissions_page(authenticated_client, real_bedrock_server):
 
 def test_configure_service_page(authenticated_client, real_bedrock_server):
     """Test the configure_service_page route with a successful response."""
+    # Updated to point to legacy route
     response = authenticated_client.get(
-        f"/server/{real_bedrock_server.server_name}/configure_service"
+        f"/legacy/server/{real_bedrock_server.server_name}/configure_service"
     )
     assert response.status_code == 200
     assert "Service" in response.text
